2024-08-12-09h52-Bilel-MigrationSrc

+++ calorie-tracker-backend/src/exercises/exercise.service.ts
@@ -3,6 +3,7 @@ import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { Exercise } from './schemas/exercise.schema';
 import { CreateExerciseDto } from './dto/create-exercise.dto';
+import { UpdateExerciseDto } from './dto/update-exercise.dto';
 
 @Injectable()
 export class ExerciseService {
@@ -11,8 +12,33 @@ export class ExerciseService {
   ) {}
 
   async create(createExerciseDto: CreateExerciseDto): Promise<Exercise> {
+    // Vérifiez si un exercice similaire existe déjà
+    const existingExercise = await this.exerciseModel.findOne({
+      userId: createExerciseDto.userId,
+      exerciseName: createExerciseDto.exerciseName,
+      date: createExerciseDto.date,
+    });
+
+    if (existingExercise && typeof existingExercise._id === 'string') {
+      // Si un exercice similaire existe et l'identifiant est une chaîne de caractères, mettez-le à jour
+      return this.update(existingExercise._id as string, createExerciseDto);
+    }
+
+    // Sinon, créez un nouvel exercice
     const exercise = new this.exerciseModel(createExerciseDto);
     return exercise.save();
+  }
+
+  async update(
+    id: string,
+    updateExerciseDto: UpdateExerciseDto,
+  ): Promise<Exercise> {
+    if (typeof id !== 'string') {
+      throw new Error('Invalid ID type');
+    }
+    return this.exerciseModel
+      .findByIdAndUpdate(id, updateExerciseDto, { new: true })
+      .exec();
   }
 
   async findAll(): Promise<Exercise[]> {
@@ -24,6 +50,9 @@ export class ExerciseService {
   }
 
   async delete(id: string): Promise<Exercise> {
+    if (typeof id !== 'string') {
+      throw new Error('Invalid ID type');
+    }
     return this.exerciseModel.findByIdAndDelete(id).exec();
   }
 }
